requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
python-multipart==0.0.6
Pillow==9.5.0
pytesseract==0.3.10
beautifulsoup4==4.12.2